Add reminders to banner
